Reorganized a lot of the cooked-mode code to make it easier to explain.
Also includes some documentation & cleanup done while writing the blog post describing this code.

Another refactoring of GETC_TTY_COOKED

Tearing out the echo buffer in another attempt to make this easier to read and talk about. Now that we call setdevice it's less cumbersome to have echo in different places, and it'll make it easier to add more complex cursor handling when I get around to it.